+++ modules/apps/portal-cache/portal-cache-multiple/src/main/java/com/liferay/portal/cache/multiple/internal/configuration/PortalCacheClusterConfiguration.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.cache.multiple.configuration;
+package com.liferay.portal.cache.multiple.internal.configuration;
 
 import aQute.bnd.annotation.metatype.Meta;
